Renaming before publishing - camp = minesattestations. add in a premlinary seed.
